Create Acquisition section
this is the section of the library that write down the materials needed by the library and when these materials have been ordered for and received, they check their list written to see if it correlates with what they ordered for, if it doesn't the materials not ordered for but received are then return back to the vendor while the ones that correlates with what was ordered would be stamped to claim ownership by the library. When new materials ordered are received, they are first taken to the acquisition section.